feat(shared): move workspace plan types into shared
