+++ packages/server/modules/gatekeeper/tests/unit/workspacePlans.spec.ts
@@ -1,10 +1,10 @@
 import { InvalidWorkspacePlanStatus } from '@/modules/gatekeeper/errors/billing'
 import { updateWorkspacePlanFactory } from '@/modules/gatekeeper/services/workspacePlans'
-import { WorkspacePlan } from '@/modules/gatekeeperCore/domain/billing'
 import { EventBusEmit } from '@/modules/shared/services/eventBus'
 import { WorkspaceNotFoundError } from '@/modules/workspaces/errors/workspace'
 import { WorkspaceWithOptionalRole } from '@/modules/workspacesCore/domain/types'
 import { expectToThrow } from '@/test/assertionHelper'
+import { WorkspacePlan } from '@speckle/shared'
 import { expect } from 'chai'
 import cryptoRandomString from 'crypto-random-string'
 import { omit } from 'lodash'
